% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adjacent.R
\name{adjacentCells}
\alias{adjacentCells}
\title{Raster adjacencies}
\usage{
adjacentCells(r, cells, directions = 8, include = FALSE, dataTable = FALSE)
}
\arguments{
\item{txt}{Text to append to "foot"}
}
\value{
TBD.
}
\description{
Find adjacent grid cells in a regular raster
}
\author{
Chris Jochem
}
