% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometry.R
\name{calc_lw}
\alias{calc_lw}
\title{Length and width calculator}
\usage{
calc_lw(X)
}
\arguments{
\item{X}{minimum bounding rectangles of building footprints of type \code{sf}.}
}
\value{
numeric vectors of length and width measured for each item in \code{X}.
}
\description{
Internal helper geometry function.
}
\keyword{internal}
