% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_footstats.R
\name{calculate_footstats}
\alias{calculate_footstats}
\alias{calculate_footstats.sf}
\alias{calculate_footstats.sfc}
\alias{calculate_footstats.sp}
\alias{calculate_footstats.character}
\alias{calculate_footstats.list}
\title{calculate_footstats: Feature statistics of building footprints}
\usage{
calculate_footstats(
  X,
  index = NULL,
  metrics = "all",
  minArea = NULL,
  maxArea = NULL,
  controlUnits = NULL,
  controlDist = NULL,
  zoneMethod = "centroid",
  verbose = FALSE
)

\method{calculate_footstats}{sf}(
  X,
  index = NULL,
  metrics = "all",
  minArea = NULL,
  maxArea = NULL,
  controlUnits = NULL,
  controlDist = NULL,
  zoneMethod = "centroid",
  verbose = FALSE
)

\method{calculate_footstats}{sfc}(
  X,
  index = NULL,
  metrics = "all",
  minArea = NULL,
  maxArea = NULL,
  controlUnits = NULL,
  controlDist = NULL,
  zoneMethod = "centroid",
  verbose = FALSE
)

\method{calculate_footstats}{sp}(
  X,
  index = NULL,
  metrics = "all",
  minArea = NULL,
  maxArea = NULL,
  controlUnits = NULL,
  controlDist = NULL,
  zoneMethod = "centroid",
  verbose = FALSE
)

\method{calculate_footstats}{character}(
  X,
  index = NULL,
  metrics = "all",
  minArea = NULL,
  maxArea = NULL,
  controlUnits = NULL,
  controlDist = NULL,
  zoneMethod = "centroid",
  verbose = FALSE
)

\method{calculate_footstats}{list}(
  X,
  index = NULL,
  metrics = "all",
  minArea = NULL,
  maxArea = NULL,
  controlUnits = NULL,
  controlDist = NULL,
  zoneMethod = "centroid",
  verbose = FALSE
)
}
\arguments{
\item{X}{object with building footprint polygons. This argument can take
multiple spatial types, including \code{sf} and \code{sp}, or a filepath
string to a file, or a list where each member provides a spatial object or
a filepath string.}

\item{index}{A spatial polygon file defining areas. Or a character or numeric
value identifying a column within \code{X} which provides a zonal index for
summarising values. Alternatively a vector of indices can be provided. If
omitted all observations with \code{X} are assumed to be within one zone.}

\item{metrics}{character vector. Names of footprint statistics in the form of
"fs_area_mean", etc. Other options include \code{ALL} or \code{NODIST} to
calculate all available metrics and all except nearest neighbour distances,
respectively.}

\item{minArea}{numeric. Minimum footprint area to filter \code{X}.}

\item{maxArea}{numeric. Maximum footprint area to filter \code{X}.}

\item{controlUnits}{(optional) named list. Elements can include
\code{areaUnit}, \code{perimUnit}, and \code{distUnit}. The values for
these items should be strings that can be coerced into a \code{units}
object.}

\item{controlDist}{(Optional) named list to override default options for
distance calculations. Elements can include \code{maxSearch} and
\code{method}. Ignored if \code{metrics} does not include a distance
calculation. See \code{\link[foot]{fs_nndist}}.}

\item{zoneMethod}{One of \code{'centroid', 'intersect', 'clip'}. How should
footprints which span zones be clipped? Default is \code{'centroid'}.
See \code{\link[foot]{zonalIndex}} for details.}

\item{verbose}{logical. Should progress messages be printed. 
Default \code{False}.}
}
\value{
a \code{data.table} with an 'index' column identify the zone and
  named columns for each footprint summary statistic.
}
\description{
Calculate groups of metrics for building footprint datasets
}
\details{
calculate_footstats
}
\examples{
data("kampala", package="foot")
buildings <- kampala$buildings
adminzones <- kampala$adminZones

calculate_footstats(buildings, 
                    index=adminzones, 
                    metrics=c("fs_area_mean","fs_area_cv"))    

}
