% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pixel_footstats.R
\name{calculate_pixel_footstats}
\alias{calculate_pixel_footstats}
\alias{calculate_pixel_footstats.sf}
\alias{calculate_pixel_footstats.character}
\title{calculate_pixel_footstats}
\usage{
calculate_pixel_footstats(
  X,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  parallel = TRUE,
  tileSize = c(1000, 1000),
  overlap = 0,
  file = NULL
)

\method{calculate_pixel_footstats}{sf}(
  X,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  parallel = TRUE,
  tileSize = c(1000, 1000),
  overlap = 0,
  file = NULL
)

\method{calculate_pixel_footstats}{character}(
  X,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  parallel = TRUE,
  tileSize = c(1000, 1000),
  overlap = 0,
  file = NULL
)
}
\arguments{
\item{txt}{Text to append to "foot"}
}
\value{
TBD.
}
\description{
Calculate selected metrics of building footprints
}
\details{
calculate_pixel_footstats
}
\author{
Chris Jochem
}
