% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{circularFilter}
\alias{circularFilter}
\alias{make_circular_filter}
\title{Circular Filter}
\usage{
make_circular_filter(size = 5)
}
\arguments{
\item{size}{The length of the side of a square matrix. The circular window
has a radius of size/2.}
}
\value{
Matrix with values inside the circle set to 1, the center set to 0,
  and corners as NA. make weights window
}
\description{
Make an approximately circular filter inside of an 'n x n'
  matrix.
}
\details{
Filters for gridded data
}
\examples{
f <- foot::make_circular_filter(11)
print(f)
class(f)


}
