% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footFun.R
\name{footFun}
\alias{footFun}
\title{Example function}
\usage{
footFun(txt = "funk")
}
\arguments{
\item{txt}{Text to append to "foot"}
}
\value{
Prints a message to the R console.
}
\description{
This is an example function that prints a message.
}
