% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_angle_entropy.R
\name{fs_angle_entropy}
\alias{fs_angle_entropy}
\alias{fs_angle_entropy.sp}
\alias{fs_angle_entropy.sf}
\title{Building angle calculation}
\usage{
fs_angle_entropy(X, index = NULL, col = NULL, normalize = TRUE)

\method{fs_angle_entropy}{sp}(X, index = NULL, col = NULL, normalize = TRUE)

\method{fs_angle_entropy}{sf}(X, index = NULL, col = NULL, normalize = TRUE)
}
\arguments{
\item{txt}{Text to append to "foot"}
}
\value{
TBD.
}
\description{
Calculate selected metrics of building footprints
}
\author{
Chris Jochem
}
