% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometry.R
\name{fs_area}
\alias{fs_area}
\title{Area}
\usage{
fs_area(X, unit = NULL)
}
\arguments{
\item{X}{polygons of building footprints of type \code{sf}.}

\item{unit}{string indicating unit of measure. Passed to
\code{units::set_units}.}
}
\value{
numeric vector of area measured for each item in \code{X}.
}
\description{
Helper geometry function to measure the area of building
  footprint polygons.
}
