% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_area_cv.R
\name{fs_area_cv}
\alias{fs_area_cv}
\alias{fs_area_cv.sp}
\alias{fs_area_cv.sf}
\title{Building area coefficient of variation (CV)}
\usage{
fs_area_cv(X, index = NULL, unit = NULL, col = NULL)

\method{fs_area_cv}{sp}(X, index = NULL, unit = NULL, col = NULL)

\method{fs_area_cv}{sf}(X, index = NULL, unit = NULL, col = NULL)
}
\arguments{
\item{X}{Spatial object with building footprint polygons}

\item{index}{A string identifying a column within \code{X} which provides a
zonal index for summarising values. Alternatively a vector of indices can
be provided. If omitted, all observations in \code{X} are assumed to be
within one zone.}

\item{unit}{character or \code{units} object to define area. Default is
\code{NULL} which will use the units of the spatial reference system}

\item{col}{column name within \code{X} with pre-calculated area measures}
}
\value{
\code{data.table} of zonal indices and values
}
\description{
Calculate and summarise selected metrics of building 
footprint polygons within zones.
}
