% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_area_max.R
\name{fs_area_max}
\alias{fs_area_max}
\alias{fs_area_max.sp}
\alias{fs_area_max.sf}
\title{Building area maximum calculation}
\usage{
fs_area_max(X, index = NULL, unit = NULL, col = NULL)

\method{fs_area_max}{sp}(X, index = NULL, unit = NULL, col = NULL)

\method{fs_area_max}{sf}(X, index = NULL, unit = NULL, col = NULL)
}
\value{
\code{data.table} of zonal indices and values
}
\description{
Calculate and summarise selected metrics of building 
footprint polygons within zones.
#' @inheritParams fs_area_mean
}
