% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_area_min.R
\name{fs_area_min}
\alias{fs_area_min}
\alias{fs_area_min.sp}
\alias{fs_area_min.sf}
\title{Building area minimum calculation}
\usage{
fs_area_min(X, index = NULL, unit = NULL, col = NULL)

\method{fs_area_min}{sp}(X, index = NULL, unit = NULL, col = NULL)

\method{fs_area_min}{sf}(X, index = NULL, unit = NULL, col = NULL)
}
\value{
\code{data.table} of zonal indices and values
}
\description{
Calculate and summarise selected metrics of building 
footprint polygons within zones.
#' @inheritParams fs_area_mean
}
