% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_area_sd.R
\name{fs_area_sd}
\alias{fs_area_sd}
\alias{fs_area_sd.sp}
\alias{fs_area_sd.sf}
\title{Building area calculation}
\usage{
fs_area_sd(X, index = NULL, unit = NULL, col = NULL)

\method{fs_area_sd}{sp}(X, index = NULL, unit = NULL, col = NULL)

\method{fs_area_sd}{sf}(X, index = NULL, unit = NULL, col = NULL)
}
\arguments{
\item{txt}{Text to append to "foot"}
}
\value{
TBD.
}
\description{
Calculate selected metrics of building footprints
}
\author{
Chris Jochem
}
