% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_area_total.R
\name{fs_area_total}
\alias{fs_area_total}
\alias{fs_area_total.sp}
\alias{fs_area_total.sf}
\title{Building area total calculation}
\usage{
fs_area_total(X, index = NULL, unit = "ha", col = NULL)

\method{fs_area_total}{sp}(X, index = NULL, unit = "ha", col = NULL)

\method{fs_area_total}{sf}(X, index = NULL, unit = "ha", col = NULL)
}
\arguments{
\item{txt}{Text to append to "foot"}
}
\value{
TBD.
}
\description{
Calculate selected metrics of building footprints
}
\author{
Chris Jochem
}
