% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_compact_median.R
\name{fs_compact_median}
\alias{fs_compact_median}
\alias{fs_compact_median.sp}
\alias{fs_compact_median.sf}
\title{Building compactness measures}
\source{
Polsby, Daniel D., and Robert D. Popper. 1991. “The Third Criterion:
  Compactness as a procedural safeguard against partisan gerrymandering.”
  Yale Law & Policy Review 9 (2): 301–353.
}
\usage{
fs_compact_median(X, index = NULL, unit = NULL, col = NULL)

\method{fs_compact_median}{sp}(X, index = NULL, unit = NULL, col = NULL)

\method{fs_compact_median}{sf}(X, index = NULL, unit = NULL, col = NULL)
}
\arguments{
\item{X}{Spatial object with building footprint polygons}

\item{index}{A character or numeric value identifying a column within \code{X} 
which provides a zonal index for summarising values. Alternatively a vector of 
indices can be provided. If omitted all observations in \code{X} are assumed 
to be within one zone.}

\item{unit}{character or \code{units} object to define area. 
Default is \code{NULL} which will use the units of the spatial reference system}

\item{col}{column name within \code{X} with pre-calculated area measures}
}
\value{
\code{data.table} of zonal indices and values
}
\description{
Calculate and summarise selected metrics of building footprint
  polygons within zones.
}
\details{
The compactness measure is the Polsby-Popper test. This is
  summarised as the mean for all footprints in the zone identified by
  \code{index}.
  
  \deqn{ PP_z = \frac{4$\pi$ * A_z}{P_z^2} },
  where P and A are the perimeter and area of zone 'z', respectively.
}
