% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_settled_count.R
\name{fs_count}
\alias{fs_count}
\alias{fs_count.sp}
\alias{fs_count.sf}
\title{Count of locations per pixel}
\usage{
fs_count(X, index = NULL, col = NULL)

\method{fs_count}{sp}(X, index = NULL, col = NULL)

\method{fs_count}{sf}(X, index = NULL, col = NULL)
}
\arguments{
\item{txt}{Text to append to "foot"}
}
\value{
TBD.
}
\description{
Calculate selected metrics of building footprints
}
\author{
Chris Jochem
}
