% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_functions.R
\name{fs_functions}
\alias{fs_functions}
\alias{binary}
\alias{count}
\alias{cv}
\alias{entropy}
\alias{majority}
\title{Built-in 'foot' functions}
\usage{
binary(x)

count(x)

cv(x)

entropy(x)

majority(x)
}
\arguments{
\item{x}{numeric vector of values to summarize.}
}
\value{
numeric value
}
\description{
In addition to generic \code{R} functions (e.g. 'mean', 'max',
  etc.), these functions are designed to provide simple access to common
  calculations.
}
\details{
These functions are designed to be used within
  \code{calculate_footstats} which processes a \code{data.table} by group ID.
  Therefore all functions take a vector of values and return a single summary
  statistic. These functions are not likely to be used on their own.
}
