% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometry.R
\name{fs_leqw}
\alias{fs_leqw}
\title{Length to equivalent-width ratio}
\usage{
fs_leqw(X, unit = NULL)
}
\arguments{
\item{X}{polygons of building footprints of type \code{sf}.}

\item{unit}{string indicating unit of measure. Passed to
\code{units::set_units}.}
}
\value{
numeric vector of length / width measured for each item in \code{X}.
}
\description{
Helper geometry function to measure calculate the ratio between
  the length and the equivalent width.
}
\details{
The length is defined as the longest side of the rotated minimum
  bounding rectangle. The equivalent width is defined as the area of the
  footprint divided by its length.
}
