% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometry.R
\name{fs_lw}
\alias{fs_lw}
\title{Length-width ratio}
\usage{
fs_lw(X)
}
\arguments{
\item{X}{polygons of building footprints of type \code{sf}.}
}
\value{
numeric vector of length / width measured for each item in \code{X}.
}
\description{
Helper geometry function to measure calculate the ratio between
  the length and width.
}
\details{
The length is defined as the longest side of the rotated minimum
  bounding rectangle and the width is the shorter side.
}
