% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometry.R
\name{fs_lwratio}
\alias{fs_lwratio}
\alias{fs_leqwratio}
\title{Length-width ratio}
\usage{
fs_lwratio(X)

fs_leqwratio(X, unit = NULL)
}
\arguments{
\item{X}{polygons of building footprints of type \code{sf}.}

\item{unit}{string indicating unit of measure. Passed to
\code{units::set_units}.}
}
\value{
numeric vector of length / width measured for each item in \code{X}.

numeric vector of length / width measured for each item in \code{X}.
}
\description{
Helper geometry function to measure calculate the ratio between
  the length and width.

Helper geometry function to measure calculate the ratio between
  the length and the equivalent width.
}
\details{
The length is defined as the longest side of the rotated minimum
  bounding rectangle and the width is the shorter side.

The length is defined as the longest side of the rotated minimum
  bounding rectangle. The equivalent width is defined as the area of the
  footprint divided by its length.
}
