% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometry.R
\name{fs_mbr}
\alias{fs_mbr}
\title{Rotated minimum bounding rectangle}
\source{
\link[https://gis.stackexchange.com/questions/22895/finding-minimum-area-rectangle-for-given-points]{https://gis.stackexchange.com/questions/22895/finding-minimum-area-rectangle-for-given-points}
}
\usage{
fs_mbr(X, returnShape = FALSE)
}
\arguments{
\item{X}{polygons of building footprints in type \code{sf}.}

\item{returnShape}{logical. Should the function return the \code{sf} polygon
of the rotated bounding rectange or should it retun the angle (in degrees).}
}
\value{
a numeric angle from 0 to 360 degrees or the rotated rectangle as a
  polygon of type \code{sf}.
}
\description{
Helper function to provide the minimum rotated bounding
  rectangle of a polygon or set of points. Optionally, the function will
  return the bearing angle in degrees from vertical in a clockwise direction.
}
\details{
This function is currently not vectorized and processing is limited
  to one shape.
}
