% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_NNdist_median.R
\name{fs_nndist_median}
\alias{fs_nndist_median}
\alias{fs_nndist_median.sp}
\alias{fs_nndist_median.sf}
\title{Nearest neighbour median distance}
\usage{
fs_nndist_median(X, index = NULL, unit = NULL, col = NULL, ...)

\method{fs_nndist_median}{sp}(X, index = NULL, unit = NULL, col = NULL, ...)

\method{fs_nndist_median}{sf}(X, index = NULL, unit = NULL, col = NULL, ...)
}
\arguments{
\item{X}{Spatial object with building footprint polygons}

\item{index}{A character or numeric value identifying a column within \code{X} 
which provides a zonal index for summarising values. Alternatively a vector of 
indices can be provided. If omitted all observations in \code{X} are assumed 
to be within one zone.}

\item{unit}{character or \code{units} object to define area. 
Default is \code{NULL} which will use the units of the spatial reference system}

\item{col}{column name within \code{X} with pre-calculated area measures}

\item{...}{Additional arguments to pass to \code{fs_nndist}.}
}
\value{
\code{data.table} of zonal indices and values.
}
\description{
Calculate and summarise selected metrics of building 
footprint polygons within zones.
}
