% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_NNdist_sd.R
\name{fs_NNdist_sd}
\alias{fs_NNdist_sd}
\alias{fs_NNdist_sd.sp}
\alias{fs_NNdist_sd.sf}
\title{Nearest neighbour mean distance}
\usage{
fs_NNdist_sd(X, index = NULL, unit = NULL, col = NULL)

\method{fs_NNdist_sd}{sp}(X, index = NULL, unit = NULL, col = NULL)

\method{fs_NNdist_sd}{sf}(X, index = NULL, unit = NULL, col = NULL)
}
\arguments{
\item{X}{Spatial object with building footprint polygons}

\item{index}{A character or numeric value identifying a column within \code{X} 
which provides a zonal index for summarising values. Alternatively a vector of 
indices can be provided. If omitted all observations in \code{X} are assumed 
to be within one zone.}

\item{unit}{character or \code{units} object to define area. 
Default is \code{NULL} which will use the units of the spatial reference system}

\item{col}{column name within \code{X} with pre-calculated area measures}
}
\value{
\code{data.table} of zonal indices and values.
}
\description{
Calculate and summarise selected metrics of building 
footprint polygons within zones.
}
\author{
Chris Jochem
}
