% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_perim_cv.R
\name{fs_perim_cv}
\alias{fs_perim_cv}
\alias{fs_perim_cv.sp}
\alias{fs_perim_cv.sf}
\title{Building perimeter coefficient of variation (CV)}
\usage{
fs_perim_cv(X, index = NULL, unit = NULL, col = NULL)

\method{fs_perim_cv}{sp}(X, index = NULL, unit = NULL, col = NULL)

\method{fs_perim_cv}{sf}(X, index = NULL, unit = NULL, col = NULL)
}
\value{
\code{data.table} of zonal indices and values
}
\description{
Calculate and summarise selected metrics of building 
footprint polygons within zones.
#' @inheritParams fs_area_mean
}
