% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_perim_mean.R
\name{fs_perim_mean}
\alias{fs_perim_mean}
\alias{fs_perim_mean.sp}
\alias{fs_perim_mean.sf}
\title{Building area calculation}
\usage{
fs_perim_mean(X, index = NULL, unit = NULL, col = NULL)

\method{fs_perim_mean}{sp}(X, index = NULL, unit = NULL, col = NULL)

\method{fs_perim_mean}{sf}(X, index = NULL, unit = NULL, col = NULL)
}
\arguments{
\item{txt}{Text to append to "foot"}
}
\value{
TBD.
}
\description{
Calculate selected metrics of building footprints
}
\author{
Chris Jochem
}
