% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_perim_median.R
\name{fs_perim_median}
\alias{fs_perim_median}
\alias{fs_perim_median.sp}
\alias{fs_perim_median.sf}
\title{Building perimeter median calculation}
\usage{
fs_perim_median(X, index = NULL, unit = NULL, col = NULL)

\method{fs_perim_median}{sp}(X, index = NULL, unit = NULL, col = NULL)

\method{fs_perim_median}{sf}(X, index = NULL, unit = NULL, col = NULL)
}
\value{
\code{data.table} of zonal indices and values
}
\description{
Calculate and summarise selected metrics of building 
footprint polygons within zones.
#' @inheritParams fs_area_mean
}
