% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_perim_sd.R
\name{fs_perim_sd}
\alias{fs_perim_sd}
\alias{fs_perim_sd.sp}
\alias{fs_perim_sd.sf}
\title{Building perimeter standard deviation calculation}
\usage{
fs_perim_sd(X, index = NULL, unit = NULL, col = NULL)

\method{fs_perim_sd}{sp}(X, index = NULL, unit = NULL, col = NULL)

\method{fs_perim_sd}{sf}(X, index = NULL, unit = NULL, col = NULL)
}
\value{
\code{data.table} of zonal indices and values
}
\description{
Calculate and summarise selected metrics of building 
footprint polygons within zones.
#' @inheritParams fs_area_mean
}
