% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_settled_binary.R
\name{fs_settled}
\alias{fs_settled}
\alias{fs_settled.sp}
\alias{fs_settled.sf}
\title{Binary settled status}
\usage{
fs_settled(X, index = NULL, col = NULL)

\method{fs_settled}{sp}(X, index = NULL, col = NULL)

\method{fs_settled}{sf}(X, index = NULL, col = NULL)
}
\arguments{
\item{txt}{Text to append to "foot"}
}
\value{
TBD.
}
\description{
Calculate selected metrics of building footprints
}
\author{
Chris Jochem
}
