% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometry.R
\name{fs_shape}
\alias{fs_shape}
\alias{fs_compact}
\title{Compactness index}
\source{
Polsby, Daniel D., and Robert D. Popper. 1991. “The Third Criterion:
  Compactness as a procedural safeguard against partisan gerrymandering.”
  Yale Law & Policy Review 9 (2): 301–353.
}
\usage{
fs_compact(X)

fs_shape(X)
}
\arguments{
\item{X}{polygons of building footprints of type \code{sf}.}
}
\value{
numeric vector of compactness values ranging from 0 to 1 for each
  item in \code{X}.

numeric vector with shape index values ranging from 0 to 1 or each
  item in \code{X}.
}
\description{
Calculates an approximate measure of shape "compactness".

Calculates a measure of shape or "roundness".
}
\details{
The compactness measure is the Polsby-Popper test based on the
  perimeter and area of each footprint.
  
  \deqn{ PP_i = \frac{4$\pi$ * A_z}{P_i^2} },
  where P and A are the perimeter and area of footprint 'i', respectively.
  Values closer to 1 indicate more compact shapes.

The shape index is calculated as the ratio of footprint polygon's
  area to the area of minimum bounding circle. The mbc is calculated using
  \code{fs_mbc} and \code{lwgeom}. Values closer 1 suggest more rounded
  shapes.

The function first looks for pre-calculated values of area in a field called
\code{fs_area}. If not present, the areas are calculated in m^2.
}
