% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_shape_median.R
\name{fs_shape_median}
\alias{fs_shape_median}
\alias{fs_shape_median.sp}
\alias{fs_shape_median.sf}
\title{Building shape measure}
\usage{
fs_shape_median(X, index = NULL, unit = NULL, col = NULL)

\method{fs_shape_median}{sp}(X, index = NULL, unit = NULL, col = NULL)

\method{fs_shape_median}{sf}(X, index = NULL, unit = NULL, col = NULL)
}
\arguments{
\item{X}{Spatial object with building footprint polygons}

\item{index}{A character or numeric value identifying a column within \code{X} 
which provides a zonal index for summarising values. Alternatively a vector of 
indices can be provided. If omitted all observations in \code{X} are assumed 
to be within one zone.}

\item{unit}{character or \code{units} object to define area. 
Default is \code{NULL} which will use the units of the spatial reference system}

\item{col}{column name within \code{X} with pre-calculated area measures}
}
\value{
\code{data.table} of zonal indices and values
}
\description{
Calculate and summarise selected metrics of building footprint
  polygons within zones.
}
\details{
The shape index is calculated as the ratio of footprint polygon's
  area to the area of minimum bounding circle. The mbc is calculated using
  \code{fs_mbc} and \code{lwgeom}. Values closer 1 suggest more rounded
  shapes.

The function first looks for pre-calculated values of area in a field called
\code{fs_area}. If not present, the areas are calculated in m^2.
}
