% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlist.R
\name{fs_varlist}
\alias{fs_varlist}
\alias{is.fs_varlist}
\title{Select variables}
\usage{
fs_varlist(...)

is.fs_varlist(x)
}
\arguments{
\item{...}{Variables to include together as parameters for summary functions.}

\item{x}{object to be tested.}
}
\description{
Functions to construct and check for lists of variables for
  \code{foot}.
}
\details{
This helper function is used by \code{calculate_footstats} to select
  multiple columns or building characteristics to be used in a summary
  function. This facilitates using functions that take multiple arguments.
}
\seealso{
\code{[calculate_footstats]} for examples of summary functions.
}
