% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_functions.R
\name{gen_nnindex}
\alias{gen_nnindex}
\title{Generate a nearest neighbour index function}
\usage{
gen_nnindex(z, zoneField = NULL, unit = "m")
}
\arguments{
\item{z}{A spatial polygon object of \code{sf} or \code{sp} type. If
omitted all observations in \code{X} are assumed to be within one zone and
the area of the minimum bounding circle is used for the nearest neighbour
index.}

\item{zoneField}{(Optional) Column name of unique identifiers in \code{zone}
to use. If omitted, the 'zoneID' will be numbered \code{1:nrow(zone)}.}

\item{unit}{character or \code{units} object to define distance. Default will
attempt to coerce units to meters.}
}
\description{
Creates a new instance of the \code{fs_nnindex} function and
  initialises it with zone and unit information.
}
\details{
This is a function factory. It creates a partial function in order
  to allow \code{fs_nnindex} to be used by the internal loop of
  \code{calculate_footstats}. This function will generally not be used on its
  own.
}
