% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_fs.R
\name{get_fs_metrics}
\alias{get_fs_metrics}
\alias{list_fs}
\alias{get_fs_units}
\title{List footprint summary metrics}
\usage{
get_fs_metrics(short_names, group = NULL)

list_fs(col_names = NULL)

get_fs_units(metrics = NULL)
}
\arguments{
\item{short_names}{Character vector of short names to lookup \code{fs_}
function names.}

\item{group}{Character vector to return a group of metrics (e.g. "area").}

\item{col_names}{Character vector to select columns from the table of
metrics.}

\item{metrics}{Character vector of \code{fs_} function names to look up the
default units.}
}
\value{
Vector or data.frame with selected footprint metric function names
  and/or units.
}
\description{
Helper functions to list footprint metric names and units
}
\details{
These convenience function access the internal data in
  \code{foot::fs_footprint_metrics}. They provide easy look-up access for the
  function names, the default units of measurement, or to return columns.
  Arguments are optional and if omitted the full set of metrics will be
  returned.
}
