% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tile.R
\name{gridTiles}
\alias{gridTiles}
\alias{gridTiles.stars}
\alias{gridTiles.raster}
\title{gridTiles}
\usage{
gridTiles(X, n = NULL, px = c(1000, 1000), overlap = 0)

\method{gridTiles}{stars}(X, n = NULL, px = c(1000, 1000), overlap = 0)

\method{gridTiles}{raster}(X, n = NULL, px = c(1000, 1000), overlap = 0)
}
\arguments{
\item{X}{The gridded (raster or Stars) object to find tiles}

\item{n}{number of tiles. Can be a vector length 1 or length 2 (rows, columns)}

\item{px}{number of pixels per side of a tile. Can be a vector of length 2 
(rows, column pixels). Ignored if n provided. Default is 1000.}

\item{overlap}{number of pixels of overlap between internal tiles}
}
\value{
TBD.
}
\description{
Split gridded data extents into subsets for parallel processing.
}
\details{
gridTiles
}
\author{
Chris Jochem
}
