% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_fs.R
\name{list_fs}
\alias{list_fs}
\title{List footprint summary metrics}
\usage{
list_fs(what = "all", how = "all")
}
\arguments{
\item{what}{Character vector of names of characteristics to look up (e.g.
"area"). Alternatively, list \code{'all'} or all characteristics except
nearest neighbour distances (\code{'nodist'}).}

\item{how}{Character vector of summary functions to look up (e.g. "mean") or
\code{'all'} available metrics.}
}
\value{
Vector or \code{data.frame} with selected footprint metric function
  names and/or units.
}
\description{
Helper function to list footprint characteristic names and
  built-in summary functions.
}
\details{
Provides an easy look-up for the built-in function names, and
  basic geometric characteristics used by \code{foot}. Supplying a "what"
  characteristics finds all built-in functions. Conversely, supplying a "how"
  function name returns are characteristics for which that summary is
  available. Arguments are optional
  and if both are omitted the full set of metrics will be returned.
}
\examples{
# get the full list of all available 
list_fs()

# get all summary functions for "area"
list_fs(what="area")

# get all characteristics relevant for entropy
list_fs(how="entropy")

}
