% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_grid.R
\name{make_templateHeader}
\alias{make_templateHeader}
\title{Construct headers for gridded data}
\usage{
make_templateHeader(
  filename,
  nrows,
  ncols,
  xmin,
  ymin,
  xmax,
  ymax,
  proj,
  datatype,
  nlayers = 1,
  interleave = "BSQ",
  nodata
)
}
\arguments{
\item{filename}{character string path to desired output file.}

\item{nrows}{number of rows in the data}

\item{ncols}{number of columns in the data}

\item{xmin, ymin, xmax, ymax}{geographic extent coordinates of the output data}

\item{proj}{Proj4String of a coordinate reference system.}

\item{nlayers}{number of layers (multiband) in the gridded data. Default is 1.}

\item{interleave}{character string for band interleaving. Currently only 'BSQ' supported.}

\item{nodata}{Value to represent no data in the output.}
}
\description{
Creates a .grd file with information on resolution and spatial extent for 
use with a binary file (.gri).
}
\details{
These gridded files are the native files used by the 
\code{raster} package. The .gri (data) and .grd (header) files must have matching filenames.
}
