% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_grid.R
\name{write_imageBinary}
\alias{write_imageBinary}
\title{Fast writing of binary files}
\usage{
write_imageBinary(data, cellNumbers = 1:length(data), filename, mapMode)
}
\arguments{
\item{data}{vector of}

\item{cellNumbers}{vector of cell indices to update with data values}

\item{filename}{character string of path to the output file}

\item{mapMode}{character string of the \code{raster} data type.}
}
\description{
\code{write_imageBinary} uses memory mapping functions to write 
gridded data to a binary output format. Binary outputs can be read natively 
but some software, or with the addition of header files, be converted to 
formats easily read by most software.
}
\details{
The function uses \code{\link[mmap]{mmap}} create and update a 
binary output file on the hard drive.

The mapMode converts the \code{raster} data type strings to a \code{CType} used 
by \code{mmap}. 
\itemize{
\item "LOG1S" \code{mmap::logi8()}
\item "INT1S" \code{mmap::int8()}
\item "INT1U" \code{mmap::uint8()}
\item "INT2S" \code{mmap::int16()}
\item "INT2U" \code{mmap::uint16()}
\item "INT4S" \code{mmap::int32()}
\item "INT4U" \code{NA}
\item "INT4S" \code{mmap::real32()}
\item "INT8S" \code{mmap::real64()}
}
}
