% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zonal_index.R
\name{zonalIndex}
\alias{zonalIndex}
\alias{zonalIndex.sf}
\alias{zonalIndex.sfc}
\alias{zonalIndex.sp}
\alias{zonalIndex.stars}
\alias{zonalIndex.Raster}
\alias{zonalIndex.character}
\title{Index buildings footprints to 'zones'}
\usage{
zonalIndex(X, zone, zoneField = NULL, method = "centroid", returnObject = TRUE)

\method{zonalIndex}{sf}(X, zone, zoneField = NULL, method = "centroid", returnObject = TRUE)

\method{zonalIndex}{sfc}(X, zone, zoneField = NULL, method = "centroid", returnObject = TRUE)

\method{zonalIndex}{sp}(X, zone, zoneField = NULL, method = "centroid", returnObject = TRUE)

\method{zonalIndex}{stars}(X, zone, zoneField = NULL, method = "centroid", returnObject = TRUE)

\method{zonalIndex}{Raster}(X, zone, zoneField = NULL, method = "centroid", returnObject = TRUE)

\method{zonalIndex}{character}(X, zone, zoneField = NULL, method = "centroid", returnObject = TRUE)
}
\arguments{
\item{X}{Spatial data (or path to file) with building footprint polygons}

\item{zone}{Spatial data (or path to file) with polygon zones or a spatial
grid (i.e. "raster")}

\item{zoneField}{(Optional) Column name of unique identifiers in \code{zone}
to use. If omitted, the 'zoneID' will be numbered \code{1:nrow(zone)}.}

\item{method}{One of \code{'centroid', 'intersect', 'clip'} to determine how
footprints are allocated to zones. See details. Default is
\code{"centroid"}.}

\item{returnObject}{Logical of whether to return an sf object of X with zonal
information. Default is \code{TRUE} which is generally preferred.}
}
\value{
'sf' object with attributes of \code{X} plus the unique zone ID or a
  \code{data.table} with the row number to the record in \code{X} matched to
  the zone IDs.
}
\description{
Find the area, grid cells, or other zone that a building polygon
  overlaps or is located in. Implements an efficient spatial join by
  intersection.
}
\details{
Zone assignments for building footprints can be done using three
  possible methods, set by the \code{method=} parameter. Defining by
  'centroid' allocates and entire building and its characteristics to the
  zone(s) which its centroid intersects. Centroids are defined by
  \code{\link[sf]{st_centroid}} which could be outside the polygon shape.
  This is the default mode. Defining zones by 'intersect' uses the geometric
  binary predicate from \code{\link[sf]{st_intersects}}. This method will
  include a whole building and its characteristics into all zones that it
  intersects. Therefore a building could appear to be "counted" twice. The
  final approach, 'clip', uses \code{\link[sf]{st_intersection}} to split
  footprints so that only that area of the polygon intersecting the zone is
  include. This method is more time consuming because the geometries are
  modified.
  
  When \code{zone} is a multi-layer \code{stars} or \code{Raster*},
  \code{zoneField} can be used to select a specific layer to define the
  areas, or \code{NULL} to use each pixel as a unique zone.
}
\examples{
data("kampala", package="foot")

buildings <- kampala$buildings
clusters <- kampala$clusters

# assign zones and return a new 'sf' object
zonalIndex(buildings, clusters)

# assign all intersecting zones
zonalIndex(buildings, clusters, method="intersect")

# return only a table of indices - note column names
zonalIndex(buildings, clusters, zoneField="Id", returnObject=FALSE)

}
