% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adjacent.R
\name{adjacentCells}
\alias{adjacentCells}
\title{Raster adjacencies}
\usage{
adjacentCells(r, cells, directions = 8, include = FALSE, dataTable = FALSE)
}
\arguments{
\item{r}{A gridded, raster object on which to search.}

\item{cells}{A vector of cell numbers in \code{r} to search for adjacencies.}

\item{directions}{The dimension to define neighbouring cells by contiguity. 
Accepted options include: 8, 4, "queen", "bishop", "rook" or a matrix. See 
details. Default is 8 which is a queen's contiguity.}

\item{include}{logical. Should the cell number of interest should be considered 
adjacent to itself? Dfault is \code{FALSE}.}

\item{dataTable}{logical. Should the processing and return value use \code{data.table}?}
}
\value{
A table with cells numbers ("from") and their adjacent ("to").
}
\description{
Find adjacent grid cells in a regular raster dataset.
}
\details{
The \code{directions} parameter defines neighbouring cells. This parameter 
follows typical contiguity measures for lattice data (e.g. "queen" or "rook"), or 
more complex adjacencies can be found by supplying a matrix. To use a matrix to define 
adjacencies, the center value should be set to zero, all adjacent cells should be 1 and
any cells to ignore should be set to \code{NA}.
}
\seealso{
\code{\link{make_circular_filter}}
}
\author{
Chris Jochem
}
