% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pixel_footstats.R
\name{calculate_bigfoot}
\alias{calculate_bigfoot}
\alias{calculate_bigfoot.sf}
\alias{calculate_bigfoot.character}
\title{calculate_bigfoot}
\usage{
calculate_bigfoot(
  X,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  parallel = TRUE,
  tileSize = c(1000, 1000),
  overlap = 0,
  file = NULL
)

\method{calculate_bigfoot}{sf}(
  X,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  parallel = TRUE,
  tileSize = c(1000, 1000),
  overlap = 0,
  file = NULL
)

\method{calculate_bigfoot}{character}(
  X,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  parallel = TRUE,
  tileSize = c(1000, 1000),
  overlap = 0,
  file = NULL
)
}
\arguments{
\item{txt}{Text to append to "foot"}
}
\value{
TBD.
}
\description{
Calculate selected metrics of building footprints for
  high-spatial resolution gridded outputs.
}
\details{
calculate_bigfoot
}
