% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_footstats.R
\name{calculate_footstats}
\alias{calculate_footstats}
\alias{calculate_footstats.sf}
\alias{calculate_footstats.sfc}
\alias{calculate_footstats.sp}
\alias{calculate_footstats.character}
\title{calculate_footstats: Feature statistics of building footprints}
\usage{
calculate_footstats(
  X,
  zone,
  what = "all",
  how = "all",
  controlZone = list(zoneName = "zoneID", method = "centroid"),
  controlUnits = list(areaUnit = "m^2", perimUnit = "m", distUnit = "m"),
  controlDist = list(maxSearch = 100, method = "centroid", unit =
    controlUnits$distUnit),
  filter = list(minArea = NULL, maxArea = NULL),
  verbose = TRUE
)

\method{calculate_footstats}{sf}(
  X,
  zone,
  what = "all",
  how = "all",
  controlZone = list(zoneName = "zoneID", method = "centroid"),
  controlUnits = list(areaUnit = "m^2", perimUnit = "m", distUnit = "m"),
  controlDist = list(maxSearch = 100, method = "centroid", unit =
    controlUnits$distUnit),
  filter = list(minArea = NULL, maxArea = NULL),
  verbose = TRUE
)

\method{calculate_footstats}{sfc}(
  X,
  zone,
  what = "all",
  how = "all",
  controlZone = list(zoneName = "zoneID", method = "centroid"),
  controlUnits = list(areaUnit = "m^2", perimUnit = "m", distUnit = "m"),
  controlDist = list(maxSearch = 100, method = "centroid", unit =
    controlUnits$distUnit),
  filter = list(minArea = NULL, maxArea = NULL),
  verbose = TRUE
)

\method{calculate_footstats}{sp}(
  X,
  zone,
  what = "all",
  how = "all",
  controlZone = list(zoneName = "zoneID", method = "centroid"),
  controlUnits = list(areaUnit = "m^2", perimUnit = "m", distUnit = "m"),
  controlDist = list(maxSearch = 100, method = "centroid", unit =
    controlUnits$distUnit),
  filter = list(minArea = NULL, maxArea = NULL),
  verbose = TRUE
)

\method{calculate_footstats}{character}(
  X,
  zone,
  what = "all",
  how = "all",
  controlZone = list(zoneName = "zoneID", method = "centroid"),
  controlUnits = list(areaUnit = "m^2", perimUnit = "m", distUnit = "m"),
  controlDist = list(maxSearch = 100, method = "centroid", unit =
    controlUnits$distUnit),
  filter = list(minArea = NULL, maxArea = NULL),
  verbose = TRUE
)
}
\arguments{
\item{X}{object with building footprint polygons. This argument can take
multiple spatial types, including \code{sf} and \code{sp}, or a filepath
string to a file.}

\item{zone}{A spatial polygon file defining areas. Or a string identifying a
column within \code{X} which provides a zonal index for summarising values.
Alternatively a vector of indices (with \code{length(index)==nrow(X)}) can
be provided. If omitted, all observations in \code{X} are assumed to be
within one zone.}

\item{what}{list of strings naming the columns or built-in geometry measures to
calculate for each footprint. Other options include \code{'all'} or
\code{'nodist'} to calculate all available characteristics and all except
nearest-neighbour distance metrics.}

\item{how}{list of strings naming functions to be used to calculate summary
statistics. The functions can be built-in functions (e.g. "mean","sd"), or
user-defined function names.}

\item{controlZone}{(optional) named list. Setting controls passed on to 
\code{\link[foot]{zonalIndex}}. Elements can include \code{zoneName} and
\code{method}.}

\item{controlUnits}{(optional) named list. Elements can include
\code{areaUnit}, \code{perimUnit}, and \code{distUnit}. The values for
these items should be strings that can be coerced into a \code{units}
object.}

\item{controlDist}{(optional) named list to override default
options for distance calculations. Elements can include \code{maxSearch}
and \code{method}. Ignored if \code{metrics} does not include a distance
calculation. See \code{\link[foot]{fs_nndist}}.}

\item{filter}{(optional) named list with \code{minArea} and \code{maxArea}.
These are numeric values to filter footprints prior to processing. Default
values are \code{NULL} and do not filter any records.}

\item{verbose}{logical. Should progress messages be printed. 
Default \code{TRUE}.}
}
\value{
a \code{data.table} with an index column identifying the zones and
  named columns for each footprint summary statistic.
}
\description{
Calculate groups of metrics for building footprint datasets
}
\details{
calculate_footstats

\code{calculate_footstats} is a wrapper function combining several
  internal functions from \code{foot}. It can calculate various geometric
  measures for each footprint polygon, including area, perimeter,
  compactness, shape, angle of rotation, and nearest neighbour distance. To
  conveniently find the list of built-in characteristics and summary metrics,
  use \code{list_fs()}.
  
  The \code{what} and \code{how} arguments are lists specifying the
  characteristics and the summary metrics to calculated, respectively. Each
  "how" function will be applied to each "what" characteristic. To apply a
  metric to only a subset of characteristics, nested lists, with groups of
  characteristics and functions can be supplied. See examples. 
  
  The \code{control} arguments are sets of lists with named arguments that
  pass on these parameters to other \code{foot} functions. These are
  optional.
}
\seealso{
zonalIndex, fs_nndist, list_fs
}
