% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_footstats.R
\name{calculate_footstats}
\alias{calculate_footstats}
\alias{calculate_footstats.sf}
\alias{calculate_footstats.sp}
\alias{calculate_footstats.character}
\alias{calculate_footstats.list}
\title{calculate_footstats}
\usage{
calculate_footstats(
  X,
  index = NULL,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  file = NULL
)

\method{calculate_footstats}{sf}(
  X,
  index = NULL,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  file = NULL
)

\method{calculate_footstats}{sp}(
  X,
  index = NULL,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  file = NULL
)

\method{calculate_footstats}{character}(
  X,
  index = NULL,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  file = NULL
)

\method{calculate_footstats}{list}(
  X,
  index = NULL,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  file = NULL
)
}
\arguments{
\item{txt}{Text to append to "foot"}
}
\value{
TBD.
}
\description{
Calculate selected metrics of building footprints
}
\details{
calculate_footstats
}
\author{
Chris Jochem
}
