% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_footstats.R
\name{calculate_footstats}
\alias{calculate_footstats}
\alias{calculate_footstats.sf}
\alias{calculate_footstats.sfc}
\alias{calculate_footstats.sp}
\alias{calculate_footstats.character}
\alias{calculate_footstats.list}
\title{calculate_footstats: Feature statistics of building footprints}
\usage{
calculate_footstats(
  X,
  index = NULL,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  outputPath = NULL,
  outputTag = NULL,
  driver = "GTiff",
  verbose = FALSE
)

\method{calculate_footstats}{sf}(
  X,
  index = NULL,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  outputPath = NULL,
  outputTag = NULL,
  driver = "GTiff",
  verbose = FALSE
)

\method{calculate_footstats}{sfc}(
  X,
  index = NULL,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  outputPath = NULL,
  outputTag = NULL,
  driver = "GTiff",
  verbose = FALSE
)

\method{calculate_footstats}{sp}(
  X,
  index = NULL,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  outputPath = NULL,
  outputTag = NULL,
  driver = "GTiff",
  verbose = FALSE
)

\method{calculate_footstats}{character}(
  X,
  index = NULL,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  outputPath = NULL,
  outputTag = NULL,
  driver = "GTiff",
  verbose = FALSE
)

\method{calculate_footstats}{list}(
  X,
  index = NULL,
  metrics = "all",
  gridded = TRUE,
  template = NULL,
  outputPath = NULL,
  outputTag = NULL,
  driver = "GTiff",
  verbose = FALSE
)
}
\arguments{
\item{X}{object with building footprint polygons. This argument can take
multiple spatial types, including \code{sf} and \code{sp}, or a fileparth
string to a building footprint files, or a list where each list member
provides a spatial object or a filepath.}

\item{index}{A character or numeric value identifying a column within
\code{X} which provides a zonal index for summarising values. Alternatively
a vector of indices can be provided. If omitted all observations with
\code{X} are assumed to be within one zone.}

\item{metrics}{character vector. Names of footprint statistics in the form of
"fs_area_mean", etc. Other options include \code{ALL} or \code{NODIST} to
calculate all available metrics and all except nearest neighbour distances,
respectively.}

\item{template}{(optional). When creating a gridded output, a supplied
\code{stars} or \code{raster} dataset to align the data.}

\item{outputPath}{(optional). When creating a gridded output, a path for the
location of the output.}

\item{outputTag}{(optional). A character string that will be added tagged to
the beginning of the output gridded files. If \code{X} is a \code{list},
then the names of the list items will be used or a vector of tags of the
same length as \code{X} should be supplied.}

\item{driver}{character. Currently supports geotiff ("GTiff").}

\item{verbose}{logical. Should progress messages be printed. 
Default \code{False}.}
}
\value{
a \code{data.table} with an 'index' column and named columns for each
  footprint statistic. Alternatively, geoTiffs of
}
\description{
Calculate groups of metrics for building footprint datasets
}
\details{
calculate_footstats
}
\author{
Chris Jochem
}
