% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_settled_count.R
\name{fs_count}
\alias{fs_count}
\alias{fs_count.sp}
\alias{fs_count.sf}
\title{Count of footprint locations per zone}
\usage{
fs_count(X, index = NULL, col = NULL)

\method{fs_count}{sp}(X, index = NULL, col = NULL)

\method{fs_count}{sf}(X, index = NULL, col = NULL)
}
\arguments{
\item{X}{Spatial object with building footprint polygons}

\item{index}{A character or numeric value identifying a column within \code{X} 
which provides a zonal index for summarising values. Alternatively a vector of 
indices can be provided. If omitted all observations in \code{X} are assumed 
to be within one zone.}

\item{col}{column name within \code{X} with pre-calculated area measures}
}
\value{
\code{data.table} of zonal indices and values.
}
\description{
Calculate and summarise selected metrics of building 
footprints within zones.
}
\author{
Chris Jochem
}
