% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometry.R
\name{fs_mbr}
\alias{fs_mbr}
\title{Rotated minimum bounding rectangle}
\source{
\link[https://gis.stackexchange.com/questions/22895/finding-minimum-area-rectangle-for-given-points]{https://gis.stackexchange.com/questions/22895/finding-minimum-area-rectangle-for-given-points}
}
\usage{
fs_mbr(X, returnShape = FALSE)
}
\arguments{
\item{X}{matrix of point coordinates in two columns or a polygon of \code{sf}
type.}

\item{returnShape}{logical. Should the function return the \code{sf} polygon
of the rotated bounding rectange or should it retun the angle (in degrees).}
}
\description{
Helper function to provide the minimum rotated bounding rectange
  of a polygon or set of points. Optionally, the function will return the
  bearing angle in degrees from vertical in a clockwise direction.
}
\details{
This function is currently not vectorized and processing is limited
  to one shape.
}
