% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_NNindex.R
\name{fs_nnindex}
\alias{fs_nnindex}
\alias{fs_nnindex.sp}
\alias{fs_nnindex.sf}
\title{Building Nearest Neighbour Index (NNI)}
\usage{
fs_nnindex(X, index = NULL, unit = NULL, col = NULL)

\method{fs_nnindex}{sp}(X, index = NULL, unit = NULL, col = NULL)

\method{fs_nnindex}{sf}(X, index = NULL, unit = NULL, col = NULL)
}
\arguments{
\item{X}{Spatial object with building footprints or their centroid locations.}

\item{index}{A spatial polygon object of \code{sf} or \code{sp} type. If
omitted all observations in \code{X} are assumed to be within one zone and
the area of the minimum bounding circle is used for the nearest neighbour
index.}

\item{unit}{character or \code{units} object to define distance. Default is
NULL}

\item{col}{column name or index within \code{X} with pre-calculated distance
measures.}
}
\value{
\code{data.table} of zonal indices and values
}
\description{
Calculate and summarise selected metrics of building footprint
  polygons within zones.
}
\details{
The nearest neighbour index (NNI) is a measure of spatial clustering. It
  compares the observed mean neighbour distances with a hypothetical maximum
  of dispersed observations given the area of the zone. Note that NNI is
  sensitive to changes in the zone area.
  
  \deqn{ NNI_z = \frac{\bar{NND_z}}{(0.5 * \sqrt{\frac{A_z}{n_z}}})}, 
  where z is the zone, A is the area, NND is the mean nearest neighbour
  distance, and n is the count.
  
  If a pre-calculated neighbour distance is not supplied, the function uses
  \code{fs_nndist} to calculate the distance between centroids of the
  building footprints within the same spatial zone indicated by \code{index}.
}
