% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_perim_mean.R
\name{fs_perim_mean}
\alias{fs_perim_mean}
\alias{fs_perim_mean.sp}
\alias{fs_perim_mean.sf}
\title{Building perimeter mean calculation}
\usage{
fs_perim_mean(X, index = NULL, unit = NULL, col = NULL)

\method{fs_perim_mean}{sp}(X, index = NULL, unit = NULL, col = NULL)

\method{fs_perim_mean}{sf}(X, index = NULL, unit = NULL, col = NULL)
}
\value{
\code{data.table} of zonal indices and values
}
\description{
Calculate and summarise selected metrics of building 
footprint polygons within zones.
#' @inheritParams fs_area_mean
}
