% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_perim_total.R
\name{fs_perim_total}
\alias{fs_perim_total}
\alias{fs_perim_total.sp}
\alias{fs_perim_total.sf}
\title{Building perimeter total calculation}
\usage{
fs_perim_total(X, index = NULL, unit = NULL, col = NULL)

\method{fs_perim_total}{sp}(X, index = NULL, unit = NULL, col = NULL)

\method{fs_perim_total}{sf}(X, index = NULL, unit = NULL, col = NULL)
}
\value{
\code{data.table} of zonal indices and values
}
\description{
Calculate and summarise selected metrics of building 
footprint polygons within zones.
#' @inheritParams fs_area_mean
}
