% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_settled_binary.R
\name{fs_settled}
\alias{fs_settled}
\alias{fs_settled.sp}
\alias{fs_settled.sf}
\title{Binary indicator of settled status}
\usage{
fs_settled(X, index = NULL, col = NULL)

\method{fs_settled}{sp}(X, index = NULL, col = NULL)

\method{fs_settled}{sf}(X, index = NULL, col = NULL)
}
\arguments{
\item{X}{Spatial object with building footprint polygons}

\item{index}{A string identifying a column within \code{X} which provides a
zonal index for summarising values. Alternatively a vector of indices can
be provided. If omitted, all observations in \code{X} are assumed to be
within one zone.}

\item{col}{column name within \code{X} with pre-calculated area measures}
}
\value{
\code{data.table} of zonal indices and values.
}
\description{
Calculate and summarise selected metrics of building 
footprint polygons within zones.
}
