% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_grid.R
\name{make_templateGrid}
\alias{make_templateGrid}
\alias{write_imageBinary}
\title{Writing gridded data to binary files}
\usage{
make_templateGrid(
  filename = NULL,
  datatype = "FLT8S",
  template = NULL,
  interleave = "BSQ",
  overwrite = FALSE
)
}
\arguments{
\item{filename}{character string path to desired output file.}

\item{template}{A filepath or a gridded data object to be used as a 
template to define the output file's resolution and positioning.}

\item{interleave}{Band interleaving. Currently only BSQ supported.}

\item{overwrite}{Should the output file be overwritten if it already exists. 
Default is \code{False}.}

\item{dataType}{character string of \code{raster} data types. Default is \code{FLT8S}}

\item{proj}{Proj4String of a coordinate reference system.}

\item{nodata}{Value to indicate no data in the output.}
}
\description{
Write data to binary files and structure as native \code{raster} files.
}
\details{
These functions provide a lightweight interface to binary writing demonstrated
in the \code{\link[spatial.tools]{}} package. The advantage of using binary files is 
the opportunity to using memory mapping (\code{mmap}) to support parallel writing.
}
\author{
Chris Jochem
}
