% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zonal_index.R
\name{zonalIndex}
\alias{zonalIndex}
\alias{zonalIndex.sf}
\alias{zonalIndex.sfc}
\alias{zonalIndex.sp}
\alias{zonalIndex.stars}
\alias{zonalIndex.raster}
\alias{zonalIndex.character}
\title{Index buildings footprints to 'zones'}
\usage{
zonalIndex(X, zone, zoneField = NULL, returnObject = TRUE, clip = FALSE)

\method{zonalIndex}{sf}(X, zone, zoneField = NULL, returnObject = TRUE, clip = FALSE)

\method{zonalIndex}{sfc}(X, zone, zoneField = NULL, returnObject = TRUE, clip = FALSE)

\method{zonalIndex}{sp}(X, zone, zoneField = NULL, returnObject = TRUE, clip = FALSE)

\method{zonalIndex}{stars}(X, zone, zoneField = NULL, returnObject = TRUE, clip = FALSE)

\method{zonalIndex}{raster}(X, zone, zoneField = NULL, returnObject = TRUE, clip = FALSE)

\method{zonalIndex}{character}(X, zone, zoneField = NULL, returnObject = TRUE, clip = FALSE)
}
\arguments{
\item{X}{Spatial data (or path to file) with building footprint polygons}

\item{zone}{Spatial data (or path to file) with polygon zones or a spatial
grid ("raster")}

\item{zoneField}{(Optional) Column name of unique identifiers in \code{zone}
to use. If omitted, the zoneID will be numbered \code{1:nrow(zone)}.}

\item{returnObject}{Logical of whether to return an sf object of X with zonal
information. Default is \code{TRUE} which is generally preferred.}

\item{clip}{Logical of whether polygons of X which span multiple zones should
be clipped to the zone. If not clipped, then whole building footprints are
linked to each zone. Default \code{FALSE}}
}
\value{
'sf' object with attributes of \code{X} plus the unique zone ID or a
  \code{data.table} with the row number to the record in \code{X} matched to
  the zone IDs.
}
\description{
Find the area, grid cells, or other zone that a building polygon
  overlaps or is located in. Implements an efficient spatial join by
  intersection.
}
\examples{
data("kampala")

buildings <- kampala$buildings
clusters <- kampala$clusters

# assign zones and return a new 'sf' object
zonalIndex(buildings, clusters)
# return only a table of indices - note column names
zonalIndex(buildings, clusters, zoneField="Id", returnObject=FALSE)

}
