% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zonal_index.R
\name{zonalIndex}
\alias{zonalIndex}
\alias{zonalIndex.sf}
\alias{zonalIndex.sfc}
\alias{zonalIndex.sp}
\alias{zonalIndex.stars}
\alias{zonalIndex.raster}
\alias{zonalIndex.character}
\title{Index buildings footprints to 'zones'}
\usage{
zonalIndex(X, zone, zoneField = NULL, returnObject = TRUE, clip = FALSE)

\method{zonalIndex}{sf}(X, zone, zoneField = NULL, returnObject = TRUE, clip = FALSE)

\method{zonalIndex}{sfc}(X, zone, zoneField = NULL, returnObject = TRUE, clip = FALSE)

\method{zonalIndex}{sp}(X, zone, zoneField = NULL, returnObject = TRUE, clip = FALSE)

\method{zonalIndex}{stars}(X, zone, zoneField = NULL, returnObject = TRUE, clip = FALSE)

\method{zonalIndex}{raster}(X, zone, zoneField = NULL, returnObject = TRUE, clip = FALSE)

\method{zonalIndex}{character}(X, zone, zoneField = NULL, returnObject = TRUE, clip = FALSE)
}
\arguments{
\item{X}{Spatial data (or path to file) with building footprint polygons}

\item{zone}{Spatial data (or path to file) with polygon zones or a spatial
grid ("raster")}

\item{zoneField}{(Optional) Unique identifier for each zone}

\item{returnObject}{Logical of whether to return an sf object of X with zonal
information. Default TRUE.}

\item{clip}{Logical of whether polygons of X which span multiple zones should
be clipped to the zone. If not clipped, then whole building footprints are
linked to each zone. Default FALSE.}
}
\value{
'sf' object with attributes of X plus the unique zone ID or a
  data.table with the ID to the records in \code{X} and the zone IDs.
}
\description{
Find the area, grid cells, or other zone that a building polygon
  overlaps or is located in
}
\author{
Chris Jochem
}
